% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkToolbox--master.R
\docType{data}
\name{animals}
\alias{animals}
\title{Animal Verbal Fluency Data}
\format{A 802x48 response matrix}
\usage{
data(animals)
}
\description{
Verbal fluency reponses for the animal category. The group variable specifies
people's level of NEO-FFI-3 Openness to Experience (low = 1, moderate = 2, high = 3).
}
\details{
A binary verbal fluency response matrix (n = 354) from
Christensen, Kenett, Cotter, Beaty, & Silvia (under review).
}
\examples{
data("animals")
}
\references{
Christensen, A.P., Kenett, Y. N., Cotter, K.N., Beaty, R. E., & Silvia, P.J. (under review).
Remotely close associations:
Openness to Experience and semantic memory structure
}
\keyword{datasets}
