% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkToolbox--master.R
\docType{data}
\name{B}
\alias{B}
\title{HEXACO Openness to Experience MaST Adjacency Matrix}
\format{A 16x16 TMFG filtered adjacency matrix}
\usage{
data(B)
}
\description{
MaST filtered association matrix (n = 802) of HEXACO's Openness to Experience
from Christensen, Cotter, & Silvia (in preparation).
Nomological Network of Openness to Experience.
}
\examples{

data(B)
}
\references{
Christensen, A.P., Cotter, K.N., Silvia, P.J. (in preparation).
Nomological network of openness to experience:
A network analysis of four openness to experience inventories.
}
\keyword{datasets}
