% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkToolbox--master.R
\name{binarize}
\alias{binarize}
\title{Binarize Network}
\usage{
binarize(A)
}
\arguments{
\item{A}{An adjacency matrix of network data}
}
\value{
Returns an adjancency matrix of 1's and 0's
}
\description{
Converts weighted adjacency matrix to a binarized adjacency matrix
}
\examples{
net<-TMFG(hex)$A

hexb<-binarize(net)
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
