% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comm.cent.R
\name{comm.cent}
\alias{comm.cent}
\title{Community Centrality}
\usage{
comm.cent(A, comm)
}
\arguments{
\item{A}{An adjacency matrix of network data}

\item{comm}{A vector or matrix corresponding to the
community each node belongs to}
}
\value{
A vector of community centrality values for each specified community in the network
(larger values suggest more central positioning)
}
\description{
Computes the community centrality measure of each
community in a network
}
\examples{
A <- TMFG(neoOpen)$A

comm <- igraph::walktrap.community(convert2igraph(abs(A)))$membership

result <- comm.cent(A, comm)

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
