% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkToolbox--master.R
\name{conn}
\alias{conn}
\title{Network Connectivity}
\usage{
conn(A)
}
\arguments{
\item{A}{An adjacency matrix of network A}
}
\value{
Returns a list of the edge weights, the mean, the standard deviation, and the sum of the edge weights in the network
}
\description{
Computes the average and standard deviation of the
}
\examples{
connectivity<-conn(A)
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
