% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor2cov.R
\name{cor2cov}
\alias{cor2cov}
\title{Convert Correlation Matrix to Covariance Matrix}
\usage{
cor2cov(cormat, data)
}
\arguments{
\item{cormat}{A correlation matrix}

\item{data}{The dataset the correlation matrix is from}
}
\value{
Returns a covariance matrix
}
\description{
Converts a correlation matrix to a covariance matrix
}
\examples{
cormat <- cor(neoOpen)

covmat <- cor2cov(cormat,neoOpen)

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
