% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkToolbox--master.R
\docType{data}
\name{data}
\alias{data}
\title{HEXACO Openness to Experience Correlation Matrix}
\format{A 16x16 correlation matrix}
\usage{
data(data)
}
\description{
Correlation matrix (n = 802) of HEXACO's Openness to Experience
from Christensen, Cotter, & Silvia (in preparation).
Nomological Network of Openness to Experience.
}
\examples{

data(data)
}
\references{
Christensen, A.P., Cotter, K.N., Silvia, P.J. (in preparation).
Nomological network of openness to experience:
A network analysis of four openness to experience inventories.
}
\keyword{datasets}
