% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.all.R
\name{desc.all}
\alias{desc.all}
\title{Dataset Descriptive Statisitcs}
\usage{
desc.all(data)
}
\arguments{
\item{data}{A matrix or data frame}
}
\value{
A data frame containing values for n (number of cases),
missing (number of missing cases), mean, sd, min, and max. normal?
will contain yes/no for whether the variable is normally distributed based
on the (\code{\link{shapiro.test}}) for the entire dataset
}
\description{
Computes mean, standard deviaiton (sd), minimum value (min),
maximum value (max),
and univariate normal statistics (normal?) for the entire dataset
}
\examples{

desc.all(neoOpen)

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
