% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{desc}
\alias{desc}
\title{Variable Descriptive Statisitcs}
\usage{
desc(data, column, histplot = TRUE)
}
\arguments{
\item{data}{A matrix or data frame}

\item{column}{Column name or number in \code{data}}

\item{histplot}{A histogram plot of the variable}
}
\value{
A data frame containing values for n (number of cases),
missing (number of missing cases), mean, sd, min, and max. normal?
will contain yes/no for whether the variable is normally distributed based
on the (\code{\link{shapiro.test}})
}
\description{
Computes mean, standard deviaiton (sd), minimum value (min),
maximum value (max),
and univariate normal statistics (normal?) for a variable
}
\examples{

desc(neoOpen,1)

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
