% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkToolbox--master.R
\docType{data}
\name{hex}
\alias{hex}
\title{HEXACO Openness to Experience Response Matrix}
\format{A 802x16 response matrix}
\usage{
data(hex)
}
\description{
A response matrix (n = 802) of HEXACO's Openness to Experience
from Christensen, Cotter, & Silvia (in preparation).
Nomological Network of Openness to Experience.
}
\examples{

data(hex)
}
\references{
Christensen, A.P., Cotter, K.N., Silvia, P.J. (in preparation).
Nomological network of openness to experience:
A network analysis of four openness to experience inventories.
}
\keyword{datasets}
