% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkToolbox--master.R
\name{impact}
\alias{impact}
\title{Node Impact}
\usage{
impact(A)
}
\arguments{
\item{A}{An adjacency matrix of network data}
}
\value{
A vector of node impact values for each node in the network
}
\description{
Computes impact measure of each node in a network
}
\examples{
A<-TMFG(hex)$A

nodeimp<-impact(A)

}
\references{
Kenett, Y. N., Kenett, D. Y., Ben-Jacob, E., & Faust, M. (2011).
Global and local features of semantic networks: Evidence from the Hebrew mental lexicon.
\emph{PloS one}, \emph{6}(8), e23912.
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
