% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkToolbox--master.R
\name{is.graphical}
\alias{is.graphical}
\title{Determines if Network is Graphical}
\usage{
is.graphical(A, data)
}
\arguments{
\item{A}{A partial correlation network (adjacency matrix)}

\item{data}{The dataset the correlation matrix is from}
}
\value{
Returns a covariance matrix
}
\description{
Tests for whether the network is graphical.
Input must be a partial correlation network.
Function assumes that partial correlations were computed from a multivariate normal distribution
}
\examples{
A <- TMFG(neoOpen, normal = TRUE, partial = TRUE)$A

is.graphical(A, neoOpen)
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
