% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isSym.R
\name{isSym}
\alias{isSym}
\title{isSymmetric Wrapper Function}
\usage{
isSym(matrix)
}
\arguments{
\item{matrix}{A matrix}
}
\description{
Checks if matrix is symmetric when \code{isSymmetric} fails
}
\examples{
A <- TMFG(neoOpen)$A

isSym(A)

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
