% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkToolbox--master.R
\name{lattnet}
\alias{lattnet}
\title{Generates a Lattice Network}
\usage{
lattnet(nodes, edges)
}
\arguments{
\item{nodes}{Number of nodes in lattice network}

\item{edges}{Number of edges in lattice network}
}
\value{
Returns an adjacency matrix of a lattice network
}
\description{
Generates a lattice network
}
\examples{

latt <- lattnet(10,27)

}
\references{
Rubinov, M., & Sporns, O. (2010). 
Complex network measures of brain connectivity: Uses and interpretations. 
\emph{Neuroimage}, \emph{52}(3), 1059-1069.
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
