% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnet.score.R
\name{lnet.score}
\alias{lnet.score}
\title{Latent Network Scores}
\usage{
lnet.score(data, A, comm = c("walktrap", "louvain"), ...)
}
\arguments{
\item{data}{Must be a dataset}

\item{A}{Network. An adjacency matrix}

\item{comm}{Vector. Can be a vector of community assignments or community detection algorithms
(\code{"walktrap"} or \code{"louvain"}) can be used to determine the number of communities.
Defaults to \code{1} community.
Set to \code{"walktrap"} for the \code{\link[igraph]{cluster_walktrap}} algortihm.
Set to \code{"louvain"} for \code{\link[NetworkToolbox]{louvain}} community detection}

\item{...}{Additional arguments for \code{\link[igraph]{cluster_walktrap}}
and \code{\link[NetworkToolbox]{louvain}} community detection algorithms}
}
\value{
Returns a list containing:

\item{scores}{The standardized latent network scores for each participant
and community (including the overall score)}

\item{commCor}{Correlations between the specified or identified communities}
}
\description{
This function computes latent network scores for
factor analysis models. Latent scores are computed based on 
each node's \code{\link[NetworkToolbox]{strength}} within each
community (i.e., factor) in the network. These values are used
as network "factor loadings" for the weights of each item. Notably,
network analysis allows nodes to load onto more than one factor.
These loadings are considered in the factor scores. In addition,
if the construct is a hierarchy (e.g., personality questionnaire;
items in facet scales in a trait domain), then an overall latent
score can be computed (see argument \code{general}). These overall
latent scores are computed using \code{\link[NetworkToolbox]{comm.close}}
as weights, which are roughly similar to general factor loadings in a
CFA model (see Christensen, Golino, & Silvia, 2019). The latent score
estimates are roughly equivalent to the Maximum Likelihood method in
\code{lavaan}'s \code{\link[lavaan]{cfa}} function. An important difference
is that the latent network scores account for cross-loadings in their
estimation of latent scores.
}
\examples{
A <- TMFG(neoOpen)$A

#one community
sumadj <- lnet.score(neoOpen, A)

#theoretical communities
knowncomm <- lnet.score(neoOpen, A,
comm = c(rep(1,8),rep(2,8),rep(3,8),rep(4,8),rep(5,8),rep(6,8)))

#walktrap communities
walkadj <- lnet.score(neoOpen, A, comm = "walktrap")

}
\references{
Christensen, A. P. (2018).
NetworkToolbox: Methods and measures for brain, cognitive, and psychometric network analysis in R.
\emph{The R Journal}, \emph{10}, 422-439.
doi: \href{https://doi.org/10.32614/RJ-2018-065}{10.32614/RJ-2018-065}

Christensen, A. P., Golino, H. F., & Silvia, P. J. (2019).
A psychometric network perspective on the measurement and assessment of personality traits.
\emph{PsyArXiv}.
doi: \href{https://doi.org/10.31234/osf.io/ktejp}{10.31234/osf.io/ktejp}
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
