% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkToolbox--master.R
\name{neuralcorrtest}
\alias{neuralcorrtest}
\title{Neural Network Correlation Test}
\usage{
neuralcorrtest(bstat, nstat)
}
\arguments{
\item{bstat}{Behavioral statistic for each participant with neural data}

\item{nstat}{A statistic vector (whole-network) or matrix (ROI) from the \emph{neuralstat} function}
}
\value{
Returns correlation test statistics for given statistics
}
\description{
Correlational test for group differences for global or local network characteristics of neural network data (\strong{only t-tests})
}
\examples{
\dontrun{
neuralarray <- convertConnBrainMat()

filteredneuralarray <- neuralnetfilter(neuralarray, method = "threshold", thres = .50)

AverageShortestPathLength <- neuralstat(filteredneuralarray, statistic = "ASPL")

Degree <- neuralstat(filteredneuralarray, statistic = "deg")

bstat <- iq

WholeNetwork_corr <- neuralcorrtest(bstat, AverageShortestPathLength)

ROI_corr <- neuralcorrtest(bstat, Degree)
}
}
\references{
Wagenmakers, E. J., Verhagen, J., & Ly, A. (2016).
How to quantify the evidence for the absence of a correlation.
\emph{Behavior Research Methods}, \emph{48}(2), 413-426.

Wetzels, R., & Wagenmakers, E. J. (2012).
A default Bayesian hypothesis test for correlations and partial correlations.
\emph{Psychonomic Bulletin & Review}, \emph{19}(6), 1057-1064.
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
