% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkToolbox--master.R
\name{neuralstat}
\alias{neuralstat}
\title{Local and Global Neural Network Characteristics}
\usage{
neuralstat(filarray, statistic, progBar = TRUE, ...)
}
\arguments{
\item{filarray}{Filtered array from \emph{neuralnetfilter} function}

\item{statistic}{A statistic to compute \strong{(see examples)}}

\item{progBar}{Should progress bar be displayed? Defaults to TRUE. Set FALSE for no progress bar}

\item{...}{Additional arguments for statistics functions}
}
\value{
Returns vector of global characteristics (rows = participants, columns = statistic) or a matrix of local characteristics (rows = ROIs, columns = participants)
}
\description{
Obtains a global or local network characteristic from neural network data
}
\examples{
\dontrun{
braindata <- convertConnBrainMat()

filteredbraindata <- neuralnetfilter(braindata, method = "threshold", thres = .50)

ClusteringCoefficient <- neuralstat(filteredbraindata, statistic = "CC")

AverageShortestPathLength <- neuralstat(filteredbraindata, statistic = "ASPL")

Modularity <- neuralstat(filteredbraindata, statistic = "Q")

Smallworldness <- neuralstat(filteredbraindata, statistic = "S")

Trasitivity <- neuralstat(filteredbraindata, statistic = "transitivity")

Connectivity <- neuralstat(filteredbraindata, statistic = "conn")

BetweennessCentrality <- neuralstat(filteredbraindata, statistic = "BC")

ClosenessCentrality <- neuralstat(filteredbraindata, statistic = "LC")

Degree <- neuralstat(filteredbraindata, statistic = "deg")

NodeStrength <- neuralstat(filteredbraindata, statistic = "str")

Communities <- neuralstat(filteredbraindata, statistic = "comm")

EigenvectorCentrality <- neuralstat(filteredbraindata, statistic = "EC")

LeverageCentrality <- neuralstat(filteredbraindata, statistic = "lev")

RandomShortestPathBC <- neuralstat(filteredbraindata, statistic = "rspbc")

HybridCentrality <- neuralstat(filteredbraindata, statistic = "hybrid")

NodeImpact <- neuralstat(filteredbraindata, statistic = "impact")
}
}
\references{
Blondel, V. D., Guillaume, J. L., Lambiotte, R., & Lefebvre, E. (2008).
Fast unfolding of communities in large networks. 
\emph{Journal of Statistical Mechanics: Theory and Experiment}, \emph{2008}(10), P10008.

Joyce, K. E., Laurienti, P. J., Burdette, J. H., & Hayasaka, S. (2010).
A new measure of centrality for brain networks. 
\emph{PLoS One}, \emph{5}(8), e12200.

Kenett, Y. N., Kenett, D. Y., Ben-Jacob, E., & Faust, M. (2011).
Global and local features of semantic networks: Evidence from the Hebrew mental lexicon.
\emph{PloS one}, \emph{6}(8), e23912.

Kivimaki, I., Lebichot, B., Saramaki, J., & Saerens, M. (2016).
Two betweenness centrality measures based on Randomized Shortest Paths.
\emph{Scientific Reports}, \emph{6}(19668), 1-15.

Pozzi, F., Di Matteo, T., & Aste, T. (2013).
Spread of risk across financial markets: Better to invest in the peripheries. 
\emph{Scientific Reports}, \emph{3}(1655), 1-7.

Rubinov, M., & Sporns, O. (2010). 
Complex network measures of brain connectivity: Uses and interpretations. 
\emph{Neuroimage}, \emph{52}(3), 1059-1069.
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
