% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.overlap.R
\name{node.overlap}
\alias{node.overlap}
\title{Detects Redundant Nodes in a Network}
\usage{
node.overlap(data = NULL, A, percentile = c(0.9, 0.95, 0.99),
  method = c("abs", "sign"), bootstrap = FALSE, iter = NULL,
  network = c("glasso", "TMFG"), ...)
}
\arguments{
\item{data}{Data is necessary when \code{bootstrap = TRUE}.
Data is also necessary for zero-order and n.s. correlation proportions and
Jaccard similarity in the output.
Input data as a matrix or data frame}

\item{A}{An adjacency matrix of network data}

\item{percentile}{Sets the number of overlapping nodes to show.
Defaults to \code{.99} or the nodes in the top 1% of redundancy.
Set lower to identify more overlapping nodes}

\item{method}{Sets whether edges should be consider with absolute
or signed weights.
Defaults to \code{"sign"}}

\item{bootstrap}{Bootstraps with resampling to estimate more
reliable results.
Defaults to \code{FALSE}.
Set to \code{TRUE} to bootstrap}

\item{iter}{Number of iterations to perform in the bootstrap}

\item{network}{Network estimation method for bootstrap.
Defaults to \code{"glasso"}}

\item{...}{Additional arguments to be passed to the network estimation
method.
See \link[qgraph]{EBICglasso} and \link[NetworkToolbox]{TMFG} for more details}
}
\value{
Produces a list containing:

\item{overlap}{A table with the nodes with the highest topological similarity
(\code{Topo. Similarity}; ordered from most to least).
Also includes the zero-order correlation between nodes (\code{Zero-order}),
proportion of non-significant correlations with other nodes (\code{n.s. Prop.}), and
Jaccard similarity (\code{Jaccard}).}

\item{items}{A matrix containing the names of items that have high overlap}
}
\description{
Identifies redundant nodes in the network based on several
measures. Computes the weighted topological overlap between
each node and every other node in the network. The weighted topological
overlap is implemented using the method from Nowick et al. (2009; see references)
and the function \link[wTO]{wTO} from the wTO package. Computes the zero-order
correlation between highly topological similar nodes. Computes the proportion
of non-significant correlations of the two nodes with all other nodes
(see \link[networktools]{goldbricker}). Computes the Jaccard similarity for the
overlap of the connections between two nodes. In sum, these indices offer an
array of similarity measures that can be used to determine node redundancy
}
\examples{

result <- node.overlap(neoOpen)

}
\references{
#goldbricker
Jones, P. (2018). networktools: tools for identifying important nodes in networks.
R package version 1.2.0.
Retrieved form https://CRAN.R-project.org/package=networktools
  
#wTO
Nowick, K., Gernat, T., Almaas, E., & Stubbs, L. (2009).
Differences in human and chimpanzee gene expression patterns define an evolving network of transcription factors in brain.
\emph{Proceedings of the National Academy of Sciences}, \emph{106}, 22358-22363.
doi: \href{https://doi.org/10.1073/pnas.0911376106}{10.1073/pnas.0911376106}
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
