% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.overlap.wrapper.R
\name{node.overlap.wrapper}
\alias{node.overlap.wrapper}
\title{Determines significant values for \code{\link[NetworkToolbox]{node.overlap}}}
\usage{
node.overlap.wrapper(vec, sd.above = 2)
}
\arguments{
\item{vec}{Numeric vector.
Bootstrapped data from \code{\link[NetworkToolbox]{node.overlap}} function}

\item{sd.above}{Numeric.
Standard deviations above the mean that should be used to determine
the significance value.
Defaults to \code{2}}
}
\value{
Returns a significance value determined from the bootstrapped
values of weighted topological overlap
}
\description{
Identifies a signficant value for nodes that have a topological
overlap over a certain standard deviation above the
mean topological overlap (TO) value.
The function bins TO values by rounding to the nearest tenth. Then, the
local maximas are derived. Typically, there will be two. First, there
are
}
\examples{

#bimodal distribution
nn <- 1e4
betas <- rbeta(nn,2,2)
sims <- c(betas[1:(nn/20)]*2+3, betas[(nn/20+1):nn]*2+1)

#distribution akin to bootstrapped weighted topological overlap
wto.dist <- na.omit(ifelse(sims<=2,NA,sims))*.10
hist(wto.dist)

#significance value of the right distribution
result <- node.overlap.wrapper(wto.dist)

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
