% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openness.R
\docType{data}
\name{openness}
\alias{openness}
\alias{openness.key}
\title{Four Inventories of Openness to Experience}
\format{
A 794 x 138 response matrix (openness) and 138 x 4 matrix (openness.key)
}
\usage{
data(openness)

data(openness.key)
}
\description{
A response matrix (\emph{n} = 794) of all four Openness to Experience
inventories from Christensen, Cotter, & Silvia (2019). The
key provides inventory, facet, and item description information
for the item labels. Note that because of NEO's copyrights the
items have been shortened and paraphrased
}
\examples{
# Loading data
data("openness")
data("openness.key")

# Change item labels
colnames(openness) <- openness.key$`Item Description`

}
\references{
Christensen, A. P., Cotter, K. N., & Silvia, P. J. (2019).
Reopening openness to experience: A network analysis of four openness to experience inventories.
\emph{Journal of Personality Assessment}, \emph{101}, 574-588.
}
\keyword{datasets}
