% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkToolbox--master.R
\name{PathLengths}
\alias{PathLengths}
\title{Path Lengths}
\usage{
PathLengths(A = A, weighted = FALSE)
}
\arguments{
\item{A}{An adjacency matrix of network data.}

\item{weighted}{Is the network weighted? Defaults to FALSE. Set to TRUE for weighted measures of ASPL, E, and D.}
}
\value{
A list of ASPL, E, and D of a network.
}
\description{
Computes average shortest path length (ASPL), eccentricity (E), and diameter (D) of a network (Weighted not coded).
}
\examples{
unweighted_PL<-PathLengths(A)
weighted_PL<-PathLengths(A,weighted=TRUE)
}
\references{
Rubinov, M., & Sporns, O. (2010). 
Complex network measures of brain connectivity: Uses and interpretations. 
Neuroimage, 52(3), 1059-1069.
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
