% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cpm.R
\name{plot.cpm}
\alias{plot.cpm}
\title{Plots CPM results}
\usage{
\method{plot}{cpm}(x, ...)
}
\arguments{
\item{x}{A \code{\link[NetworkToolbox]{cpm}} object}

\item{...}{Additional arguments for \link{plot}}
}
\description{
Plots CPM results
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
