% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkToolbox--master.R
\name{prepboot}
\alias{prepboot}
\title{Bootstrapped Network Preprocessing}
\usage{
prepboot(data, method, binary = FALSE, iter = 1000, a = 0.05)
}
\arguments{
\item{data}{A set of data}

\item{method}{A network filtering method}

\item{binary}{Is dataset dichotomous? Defaults to FALSE. Set TRUE if dataset is dichotomous (tetrachoric correlations are computed)}

\item{iter}{Number of bootstrap iterations. Defaults to 1000 iterations}

\item{a}{Alpha to be used for determining the critical value of correlation coefficients}
}
\value{
A correlation matrix of the mean bootstrapped network
}
\description{
Bootstraps the sample to identify the most stable correlations
}
\examples{
\dontrun{

prepTMFG<-prepboot(data,method="TMFG")
}
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
