% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkToolbox--master.R
\docType{data}
\name{restOpen}
\alias{restOpen}
\title{Resting-state Brain Data}
\format{restOpen (268 x 268 x 144 array)}
\usage{
data(restOpen)
}
\description{
Resting-state brain data associated with NEO-PI-3's Openness to Experience (\emph{n} = 144).
}
\details{
A correlation connectivity array of brain data (rmats).
Shen et al.'s (2013) brain atlas is used.
}
\examples{
data(restOpen)
}
\references{
Beaty, R. E., Chen, Q., Christensen, A. P., Qiu, J., Silvia, P. J., & Schacter, D. L. (2018).
Brain networks of the imaginative mind: Dynamic functional connectivity of default and cognitive control networks relates to Openness to Experience.
\emph{Human Brain Mapping}, \emph{39}(2), 811-821.

Beaty, R. E., Kenett, Y. N., Christensen, A. P., Rosenberg, M. D., Benedek, M., Chen, Q., ... & Silvia, P. J. (2018).
Robust prediction of individual creative ability from brain functional connectivity.
\emph{Proceedings of the National Academy of Sciences}, 201713532.

Shen, X., Tokoglu, F., Papademetris, X., & Constable, R. T. (2013).
Groupwise whole-brain parcellation from resting-state fMRI data for network node identification.
\emph{Neuroimage}, \emph{82}, 403-415.
}
\keyword{datasets}
