% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkToolbox--master.R
\name{rmse}
\alias{rmse}
\title{Root Mean Square Error}
\usage{
rmse(base, model, test)
}
\arguments{
\item{base}{Base (or full) model to be evaulated against}

\item{model}{Reduced (or testing) model (e.g., a sparse correlation or covariance matrix)}

\item{test}{Data that is not the base dataset (i.e., comparing a training model to test data)}
}
\value{
RMSE value. Lower values suggest more similarity between the full and sparse model
}
\description{
Computes the root mean square error of a sparse model to a full model
}
\examples{
A <- solve(LoGo(hex))

root <- rmse(hex, A)

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
