% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkToolbox--master.R
\name{semnetboot}
\alias{semnetboot}
\title{Partial Bootstrapped Semantic Network Analysis}
\usage{
semnetboot(data, method = c("TMFG", "LoGo", "MaST", "threshold"),
  normal = FALSE, nodes, iter = 1000, na.data = c("pairwise", "listwise",
  "fiml", "none"), ...)
}
\arguments{
\item{data}{A set of data}

\item{method}{A network filtering method.
Defaults to "TMFG"}

\item{normal}{Should data be transformed to a normal distribution?
Defaults to FALSE.
Data is not transformed to be normal.
Set to TRUE if data should be transformed to be normal
(computes correlations using the \link[qgraph]{cor_auto} function)}

\item{nodes}{Number of nodes (i.e., variables) to use in the bootstrap.
Defaults to 50% of nodes in the network.
Otherwise accepts the number of the nodes to be included}

\item{iter}{Number of bootstrap iterations.
Defaults to 1000 iterations}

\item{na.data}{How should missing data be handled?
For "listwise" deletion the \code{\link{na.omit}} function is applied.
Set to "fiml" for Full Information Maxmimum Likelihood (\link[psych]{corFiml}).
Full Information Maxmimum Likelihood is \strong{recommended} but time consuming}

\item{...}{Additional arguments for filtering methods}
}
\value{
Returns a list that includes the original semantic network measures (origmeas; ASPL, CC, Q, S),
and the bootstrapped semantic network measures (bootmeas)
}
\description{
Bootstraps (without replacement) the nodes in the network and computes global network characteristics
}
\examples{
\dontrun{
lowO <- subset(animals, Group==1)[-1]

semTMFG<-semnetboot(lowO)

semLoGo<-semnetboot(lowO,method="LoGo")

semMaST<-semnetboot(lowO,method="MaST")

semThreshold<-semnetboot(lowO,method="threshold")
}
}
\references{
Kenett, Y. N., Wechsler-Kashi, D., Kenett, D. Y., Schwartz, R. G., Ben Jacob, E., & Faust, M. (2013).
Semantic organization in children with cochlear implants: Computational analysis of verbal fluency.
\emph{Frontiers in Psychology}, {4}(543), 1-11.
}
