% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkToolbox--master.R
\name{semnetcleaner}
\alias{semnetcleaner}
\title{Semantic Network Cleaner}
\usage{
semnetcleaner(data = data)
}
\arguments{
\item{data}{A dataset of verbal fluency or linguistic data.}
}
\value{
A binary matrix of responses (rows = participants, cols = responses)
}
\description{
An automated cleaning function for semantic network data (still in the alpha stage).
}
\examples{
responsematrix<-semnetcleaner(data)
}
\references{
qdap package: Hornik, K., & Murdoch, D. (2010).
Watch Your Spelling!.
The R Journal, 3(2), 22-28.
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
