% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkToolbox--master.R
\name{semnetmeas}
\alias{semnetmeas}
\title{Semantic Network Measures}
\usage{
semnetmeas(A)
}
\arguments{
\item{A}{An adjacency matrix of network A}
}
\value{
Returns a values for ASPL, CC, Q, and S
}
\description{
Computes the average shortest path length (ASPL), clustering coefficient(CC),
modularity (Q), and small-worldness (S)
}
\examples{
A<-TMFG(hex)$A

connectivity<-semnetmeas(A)
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
