% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkToolbox--master.R
\name{smallworldness}
\alias{smallworldness}
\title{Small-worldness Measure}
\usage{
smallworldness(A, iter = 100)
}
\arguments{
\item{A}{An adjacency matrix of network data}

\item{iter}{Number of random networks to generate, which are used to calculate the mean random ASPL and CC}
}
\value{
Returns a value of small-worldness
}
\description{
Computes the small-worldness measure of a network
}
\examples{

A<-TMFG(hex)$A

swm <- smallworldness(A)

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
