% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkToolbox--master.R
\name{splitsamp}
\alias{splitsamp}
\title{Split sample}
\usage{
splitsamp(data, samples = 10, splits = 2)
}
\arguments{
\item{data}{Must be a dataset}

\item{samples}{Number of samples to produce (Defaults to 10)}

\item{splits}{Number to divide the sample by (Defaults to 2; i.e., split-half)}
}
\value{
A list containing the training (trainSample) and testing (testSample) sizes
(training sample is made to have slightly larger sizes in the case of uneven splits),
and their respective sample sizes (trainSize and testSize)
}
\description{
Randomly splits sample into equally divded sub-samples
}
\examples{
splithalf<-splitsamp(neoOpen)
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
