% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strength.dim.R
\name{strength.dim}
\alias{strength.dim}
\title{Node Strength by Dimension}
\usage{
strength.dim(A, comm, rm.zero = FALSE)
}
\arguments{
\item{A}{An adjacency matrix of network data}

\item{comm}{A vector of community assignments}

\item{rm.zero}{Should zeros be removed from the resulting matrix?
Defaults to \code{FALSE}.
Set to \code{TRUE} to reduce the noise in the results}
}
\value{
Returns a matrix of the unstandardized within- and between-community
strength values for each node
}
\description{
Computes the within- and between-community strength of each item
for each community
}
\examples{
A <- TMFG(neoOpen)$A

wc <- louvain(A)$community

str.dim <- strength.dim(A, wc, rm.zero = TRUE)

}
\author{
Alexander P. Christensen <alexpaulchristensen@gmail.com> and Hudson F. Golino <hfg9s at virginia.edu>
}
