% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkToolbox--master.R
\name{walkboot}
\alias{walkboot}
\title{Bootstrapped Walktrap Reliability}
\usage{
walkboot(data, binary = FALSE, n = nrow(data), iter = 1000)
}
\arguments{
\item{data}{A set of data}

\item{binary}{Is dataset dichotomous? Defaults to FALSE. Set TRUE if dataset is dichotomous (tetrachoric correlations are computed)}

\item{n}{Number of people to use in the bootstrap. Defaults to full sample size}

\item{iter}{Number of bootstrap iterations. Defaults to 1000 iterations}
}
\value{
Returns a list that includes a vector of the number of communities in each bootstrapped
sample (communityvector) and proportion of number of factors found across bootstrapped samples
(reliabilities)
}
\description{
Bootstraps the sample with replace to compute walktrap reliability (TMFG-filtered networks only)
}
\examples{
\dontrun{

walkTMFG<-walkboot(hex)
}
}
\references{
Csardi, G., & Nepusz, T. (2006).
The igraph software package for complex network research.
\emph{InterJournal, Complex Systems}, \emph{1695}(5), 1-9.
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
