% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavOpen.R
\docType{data}
\name{behavOpen}
\alias{behavOpen}
\title{NEO-PI-3 for Resting-state Data}
\format{behavOpen (vector, length = 144)}
\usage{
data(behavOpen)
}
\description{
NEO-PI-3 Openness to Experience associated with resting-state data (\emph{n} = 144).
}
\details{
Behavioral data of NEO-PI-3 associated with each connectivity matrix (open).

To access the resting-state brain data, please go to \url{https://drive.google.com/file/d/1ugwi7nRrlHQYuGPzEB4wYzsizFrIMvKR/view}
}
\examples{
data("behavOpen")
}
\references{
Beaty, R. E., Chen, Q., Christensen, A. P., Qiu, J., Silvia, P. J., & Schacter, D. L. (2018).
Brain networks of the imaginative mind: Dynamic functional connectivity of default and cognitive control networks relates to Openness to Experience.
\emph{Human Brain Mapping}, \emph{39}(2), 811-821.
doi: \href{https://doi.org/10.1002/hbm.23884}{10.1002/hbm.23884}

Beaty, R. E., Kenett, Y. N., Christensen, A. P., Rosenberg, M. D., Benedek, M., Chen, Q., ... & Silvia, P. J. (2018).
Robust prediction of individual creative ability from brain functional connectivity.
\emph{Proceedings of the National Academy of Sciences}, 201713532.
doi: \href{https://doi.org/10.1073/pnas.1713532115}{10.1073/pnas.1713532115}
}
\keyword{datasets}
