% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkToolbox--master.R
\name{bootgen}
\alias{bootgen}
\title{Bootstrapped Network Generalization}
\usage{
bootgen(data, method = c("TMFG", "LoGo", "MaST", "ECOplusMaST", "ECO",
  "threshold"), n = nrow(data), iter = 1000, binary = FALSE,
  depend = FALSE, ...)
}
\arguments{
\item{data}{A set of data (missing data is handled using the Full Information Maximum Likelihood function \emph{corFiml} in the \emph{psych} package)}

\item{method}{A network filtering method. Defaults to "TMFG"}

\item{n}{Number of people to use in the bootstrap. Defaults to full sample size}

\item{iter}{Number of bootstrap iterations. Defaults to 1000 iterations}

\item{binary}{Is dataset dichotomous? Defaults to FALSE. Set TRUE if dataset is dichotomous (tetrachoric correlations are computed)}

\item{depend}{Is network a dependency (or directed) network? Defaults to FALSE. Set TRUE to generate a TMFG-filtered dependency network}

\item{...}{Additional arguments for filtering methods}
}
\value{
Returns a list that includes the original filtered network (orignet),
correlation matrix of the mean bootstrapped network (bootmat),
reliabilities of the connections in the original network (netrel),
reliabilities of the connections in the bootstrapped network (bootrel),
a plot of the bootrel reliability matrix (netrel; upper triangle = actual network reliabilites, bootrel; lower triangle = overall network reliablities),
a plot of included correlations on their reliability (ConR)
}
\description{
Bootstraps the sample to identify the most stable correlations
}
\examples{
\dontrun{
prepTMFG<-bootgen(hex)

prepLoGo<-bootgen(hex,method="LoGo")

prepMaST<-bootgen(hex,method="MaST")

prepECO<-bootgen(hex,method="ECO")

prepECOplusMaST<-bootgen(hex,method="ECOplusMaST")

prepThreshold<-bootgen(hex,method="threshold")
}
}
\references{
Tumminello, M., Coronnello, C., Lillo, F., Micciche, S., & Mantegna, R. N. (2007).
Spanning trees and bootstrap reliability estimation in correlation-based networks.
\emph{International Journal of Bifurcation and Chaos}, \emph{17}(7), 2319-2329.

Wagenmakers, E. J., Verhagen, J., & Ly, A. (2016).
How to quantify the evidence for the absence of a correlation.
\emph{Behavior Research Methods}, \emph{48}(2), 413-426.

Wei, T. & Simko, V.(2017).
R package "corrplot": Visualization of a correlation matrix (Version 0.84).
Available from \url{https://github.com/taiyun/corrplot}

Wetzels, R., & Wagenmakers, E. J. (2012).
A default Bayesian hypothesis test for correlations and partial correlations.
\emph{Psychonomic Bulletin & Review}, \emph{19}(6), 1057-1064.
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
