% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustcoeff.R
\name{clustcoeff}
\alias{clustcoeff}
\title{Clustering Coefficient}
\usage{
clustcoeff(A, weighted = FALSE)
}
\arguments{
\item{A}{An adjacency matrix of network data}

\item{weighted}{Is the network weighted?
Defaults to \code{FALSE}.
Set to \code{TRUE} for weighted measures of CC and CCi}
}
\value{
Returns a list containing:

\item{CC}{Global clustering coefficient. The average clustering coefficient for each node in the network}

\item{CCi}{Local clustering coefficient. The clustering coefficient for each node in the network}
}
\description{
Computes global clustering coefficient (CC) and local clustering coefficient (CCi)
}
\examples{
# Pearson's correlation only for CRAN checks
A <- TMFG(neoOpen, normal = FALSE)$A

#Unweighted CC
CCu <- clustcoeff(A)

#Weighted CC
CCw <- clustcoeff(A, weighted=TRUE)

}
\references{
Rubinov, M., & Sporns, O. (2010). 
Complex network measures of brain connectivity: Uses and interpretations. 
\emph{NeuroImage}, \emph{52}, 1059-1069.
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
