% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comcat.R
\name{comcat}
\alias{comcat}
\title{Communicating Nodes}
\usage{
comcat(A, comm = c("walktrap", "louvain"), cent = c("strength",
  "degree"), metric = c("across", "each"), ...)
}
\arguments{
\item{A}{An adjacency matrix of network data}

\item{comm}{Can be a vector of community assignments or community detection algorithms
("walktrap" or "louvain") can be used to determine the number of factors.
Defaults to "walktrap".
Set to "louvain" for louvain community detection}

\item{cent}{Centrality measure to be used.
Defaults to "strength".}

\item{metric}{Whether the metric should be compute for across the communities
(a single value) or for each community (a value for each community).
Defaults to "across".
Set to "each" for values for each community}

\item{...}{Additional arguments for community detection algorithms}
}
\value{
A vector containing the between-community strength value for each node
}
\description{
Computes the between-community strength for each node in the network
}
\examples{
A <- TMFG(neoOpen)$A

communicating <- comcat(A, comm = "walktrap", cent = "strength", metric = "across")

}
\references{
Blanken, T. F., Deserno, M. K., Dalege, J., Borsboom, D., Blanken, P., Kerkhof, G. A., & Cramer, A. O. (2018).
The role of stabilizing and communicating symptoms given overlapping communities in psychopathology networks.
\emph{Scientific Reports}, \emph{8}, 5854.
doi: \href{https://doi.org/10.1038/s41598-018-24224-2}{10.1038/s41598-018-24224-2}
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
