% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn.R
\name{conn}
\alias{conn}
\title{Network Connectivity}
\usage{
conn(A)
}
\arguments{
\item{A}{An adjacency matrix of a network}
}
\value{
Returns a list containing:

\item{weights}{Each edge weight in the network}

\item{mean}{The mean of the edge weights in the network}

\item{sd}{The standard deviation of the edge weights in the network}

\item{total}{The sum total of the edge weights in the network}
}
\description{
Computes the average and standard deviation of the weights in the network
}
\examples{
# Pearson's correlation only for CRAN checks
A <- TMFG(neoOpen, normal = FALSE)$A

connectivity <- conn(A)

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
