% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkToolbox--master.R
\name{cpmPredict}
\alias{cpmPredict}
\title{Connectome-based Predictive Modeling--Internal Validation}
\usage{
cpmPredict(neuralarray, bstat, thresh = 0.01, method = c("mean", "sum"),
  progBar = TRUE)
}
\arguments{
\item{neuralarray}{Array from \emph{convertConnBrainMat} function}

\item{bstat}{Behavioral statistic for each participant with neural data (a vector)}

\item{thresh}{Sets an \strong{alpha} threshold for edge weights to be retained. Defaults to .01}

\item{method}{Use "mean" or "sum" of edge strengths in the positive and negative connectomes. Defaults to "mean"}

\item{progBar}{Should progress bar be displayed? Defaults to TRUE. Set FALSE for no progress bar}
}
\value{
Returns a list containing a matrix (r coefficient (r), p-value (p-value), Bayes Factor (BF), r-squared (r-squared), root mean square error (rmse)). The list also contains the positive (posMask) and negative (negMask) masks used
}
\description{
Applies the Connectome-based Predictive Modeling approach to neural data. \strong{Please see and cite references}
}
\references{
Finn, E. S., Shen, X., Scheinost, D., Rosenberg, M. D., Huang, J., Chun, M. M., Papademetris, X., Constable, R. T. (2015).
Functional connectome fingerprinting: Identifying individuals using patterns of brain connectivity.
\emph{Nature Neuroscience}, \emph{18}(11), 1664-1671.

Rosenberg, M. D., Finn, E. S., Scheinost, D., Papademetris, X., Shen, X., Constable, R. T., Chun, M. M. (2016).
A neuromarker of sustained attention from whole-brain functional connectivity.
\emph{Nature Neuroscience}, \emph{19}(1), 165-171.

Shen, X. Finn, E. S., Scheinost, D., Rosenberg, M. D., Chun, M. M., Papademetris, X., Constable, R. T. (2017).
Using connectome-based predictive modeling to predict individual behavior from brain connectivity.
\emph{Nature Protocols}, \emph{12}(3), 506-518.

Wagenmakers, E. J., Verhagen, J., & Ly, A. (2016).
How to quantify the evidence for the absence of a correlation.
\emph{Behavior Research Methods}, \emph{48}(2), 413-426.

Wetzels, R., & Wagenmakers, E. J. (2012).
A default Bayesian hypothesis test for correlations and partial correlations.
\emph{Psychonomic Bulletin & Review}, \emph{19}(6), 1057-1064.
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
