% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkToolbox--master.R
\name{Degree}
\alias{Degree}
\title{Degree}
\usage{
Degree(A)
}
\arguments{
\item{A}{An adjacency matrix of network data}
}
\value{
A vector of degree values for each node in the network
}
\description{
Computes degree of each node in a network
}
\examples{
Deg<-Degree(A)
}
\references{
Rubinov, M., & Sporns, O. (2010). 
Complex network measures of brain connectivity: Uses and interpretations. 
Neuroimage, 52(3), 1059-1069.
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
