% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.all.R
\name{desc.all}
\alias{desc.all}
\title{Dataset Descriptive Statistics}
\usage{
desc.all(data)
}
\arguments{
\item{data}{A matrix or data frame}
}
\value{
A data frame containing values for \code{n} (number of cases),
\code{missing} (number of missing cases), \code{mean}, \code{sd}, \code{min}, and \code{max}. \code{normal?}
will contain yes/no for whether the variable is normally distributed based
on the \code{\link{shapiro.test}} for the entire dataset
}
\description{
Computes \code{mean}, standard deviation (\code{sd}), minimum value (\code{min}),
maximum value (\code{max}),
and univariate normal statistics (\code{normal?}) for the entire dataset
}
\examples{

desc.all(neoOpen)

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
