% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{desc}
\alias{desc}
\title{Variable Descriptive Statistics}
\usage{
desc(data, column, histplot = TRUE)
}
\arguments{
\item{data}{A matrix or data frame}

\item{column}{Column name or number in \code{data}}

\item{histplot}{A histogram plot of the variable}
}
\value{
A data frame containing values for \code{n} (number of cases),
\code{missing} (number of missing cases), \code{mean}, \code{sd}, \code{min}, and \code{max}. \code{normal?}
will contain yes/no for whether the variable is normally distributed based
on the \code{\link{shapiro.test}} for a variable
}
\description{
Computes \code{mean}, standard deviation (\code{sd}), minimum value (\code{min}),
maximum value (\code{max}),
and univariate normal statistics (\code{normal?}) for a variable
}
\examples{

desc(neoOpen,1)

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
