% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkToolbox--master.R
\name{ECO}
\alias{ECO}
\title{ECO Neural Network Filter}
\usage{
ECO(data, weighted = TRUE, directed = FALSE)
}
\arguments{
\item{data}{Can be a dataset or a correlation matrix}

\item{weighted}{Should network be weighted?
Defaults to TRUE.
Set to FALSE to produce an unweighted (binary) network}

\item{directed}{Is the network directed?
Defaults to FALSE.
Set TRUE if the network is directed}
}
\value{
A sparse association matrix
}
\description{
Applies the ECO neural network filtering method
}
\examples{
\dontrun{
weighted_undirected_ECOnetwork<-ECO(neoOpen)

unweighted_undirected_ECOnetwork<-ECO(neoOpen,weighted=FALSE)

weighted_directed_ECOnetwork<-ECO(neoOpen,directed=TRUE)

unweighted_directed_ECOnetwork<-ECO(neoOpen,weighted=FALSE,directed=TRUE)
}
}
\references{
Fallani, F. D. V., Latora, V., & Chavez, M. (2017).
A topological criterion for filtering information in complex brain networks.
\emph{PLoS Computational Biology}, \emph{13}(1), e1005305.
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
