% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECO.R
\name{ECO}
\alias{ECO}
\title{ECO Neural Network Filter}
\usage{
ECO(data, directed = FALSE)
}
\arguments{
\item{data}{Can be a dataset or a correlation matrix}

\item{directed}{Is the network directed?
Defaults to FALSE.
Set TRUE if the network is directed}
}
\value{
A sparse association matrix
}
\description{
Applies the ECO neural network filtering method
}
\examples{
eco.net <- ECO(neoOpen)

}
\references{
Fallani, F. D. V., Latora, V., & Chavez, M. (2017).
A topological criterion for filtering information in complex brain networks.
\emph{PLoS Computational Biology}, \emph{13}, e1005305.
doi: \href{https://doi.org/10.1371/journal.pcbi.1005305}{10.1371/journal.pcbi.1005305}
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
