% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECOplusMaST.R
\name{ECOplusMaST}
\alias{ECOplusMaST}
\title{ECO+MaST Network Filter}
\usage{
ECOplusMaST(data)
}
\arguments{
\item{data}{Can be a dataset or a correlation matrix}
}
\value{
A sparse association matrix
}
\description{
Applies the \link[NetworkToolbox]{ECO} neural network filtering method
combined with the \link[NetworkToolbox]{MaST} filtering method
}
\examples{
ECOplusMaST.net <- ECOplusMaST(neoOpen)

}
\references{
Fallani, F. D. V., Latora, V., & Chavez, M. (2017).
A topological criterion for filtering information in complex brain networks.
\emph{PLoS Computational Biology}, \emph{13}, e1005305.
doi: \href{https://doi.org/10.1371/journal.pcbi.1005305}{10.1371/journal.pcbi.1005305}
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
