% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkToolbox--master.R
\name{edgerep}
\alias{edgerep}
\title{Edge Replication}
\usage{
edgerep(A, B)
}
\arguments{
\item{A}{An adjacency matrix of network A}

\item{B}{An adjacency matrix of network B}
}
\value{
Returns the number of edges that replicate
}
\description{
Computes the number of edges that replicate between two cross-sectional networks
}
\examples{
numberofedges<-edgerep(A,B)
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
