% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkToolbox--master.R
\docType{data}
\name{hexb}
\alias{hexb}
\title{HEXACO Openness to Experience Response Matrix (Binarized)}
\format{A 802x16 response matrix}
\usage{
data(hexb)
}
\description{
A response matrix (n = 802) of HEXACO's Openness to Experience
from Christensen, Cotter, & Silvia (under review).
}
\examples{

data(hexb)
}
\references{
Christensen, A.P., Cotter, K.N., Silvia, P.J. (under review).
Nomological network of Openness to Experience:
A network analysis of four Openness to Experience inventories.
\url{http://doi.org/10.17605/OSF.IO/954A7}
}
\keyword{datasets}
