% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkToolbox--master.R
\name{Hybrid}
\alias{Hybrid}
\title{Hybrid Centrality}
\usage{
Hybrid(A)
}
\arguments{
\item{A}{An adjacency matrix of network data}
}
\value{
A vector of hybrid centrality values for each node in the network (higher values are more central, lower values are more peripheral)
}
\description{
Computes hybrid centrality of each node in a network
}
\examples{
\dontrun{

HC<-Hybrid(A)
}
}
\references{
Pozzi, F., Di Matteo, T., & Aste, T. (2013).
Spread of risk across financial markets: Better to invest in the peripheries. 
Scientific Reports, 3(1655), 1-7.
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
