% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nams.R
\name{nams}
\alias{nams}
\title{Network Adjusted Mean/Sum}
\usage{
nams(data, A, adjusted = c("mean", "sum"), comm = c("walktrap",
  "louvain"), ...)
}
\arguments{
\item{data}{Must be a dataset}

\item{A}{Adjacency matrix that has already been filtered}

\item{adjusted}{Should adjusted values be the mean or sum score?
Defaults to \code{"mean"}.
Set to \code{"sum"} for sum scores}

\item{comm}{Can be a vector of community assignments or community detection algorithms
(\code{"walktrap"} or \code{"louvain"}) can be used to determine the number of communities.
Defaults to \code{1} community.
Set to \code{"walktrap"} for the \code{\link[igraph]{cluster_walktrap}} algortihm.
Set to \code{"louvain"} for \code{\link[NetworkToolbox]{louvain}} community detection}

\item{...}{Additional arguments for \code{\link[igraph]{cluster_walktrap}}
and \code{\link[NetworkToolbox]{louvain}} community detection algorithms}
}
\value{
Returns a list containing:

\item{Standardized}{The standardized network adjusted score for each participant}

\item{Unstandardized}{The unstandardized network adjusted score (mean or sum) for each participant}

\item{CommItems}{The items associated with the specified or identified communities}

\item{CommCor}{Correlations between the specified or identified communities}
}
\description{
The \code{\link[NetworkToolbox]{hybrid}} centrality is used to
adjust the mean or sum score of participant's community scores based on each
node's centrality. Each participant's response values are multipled by
the corresponding \code{\link[NetworkToolbox]{hybrid}}
centrality value (uses "random" for BC argument).
In this way, more central nodes contribute a greater score
and less central nodes contribute a lesser score.
See Christensen (in press) for more details
}
\examples{
A <- TMFG(neoOpen)$A

#one community
sumadj <- nams(neoOpen, A, adjusted = "sum")

#theoretical communities
knowncomm <- nams(neoOpen, A,
comm = c(rep(1,8),rep(2,8),rep(3,8),rep(4,8),rep(5,8),rep(6,8)))

#walktrap communities
walkadj <- nams(neoOpen, A, adjusted = "sum", comm = "walktrap")

}
\references{
Christensen, A. P. (2018).
NetworkToolbox: Methods and measures for brain, cognitive, and psychometric network analysis in R.
\emph{PsyArXiv}.
doi: \href{https://doi.org/10.31234/osf.io/6kmav}{10.31234/osf.io/6kmav}
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
