% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoOpen.R
\docType{data}
\name{neoOpen}
\alias{neoOpen}
\title{NEO-PI-3 Openness to Experience Data}
\format{A 802x48 response matrix}
\usage{
data(neoOpen)
}
\description{
A response matrix (\emph{n} = 802) of NEO-PI-3's Openness to Experience
from Christensen, Cotter, & Silvia (2019).
}
\examples{
data("neoOpen")

}
\references{
Christensen, A. P., Cotter, K. N., & Silvia, P. J. (2019).
Reopening openness to experience: A network analysis of four openness to experience inventories.
\emph{Journal of Personality Assessment}, \emph{101}, 574-588.
}
\keyword{datasets}
