% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoOpen.R
\docType{data}
\name{neoOpen}
\alias{neoOpen}
\title{NEO-PI-3 Openness to Experience Data}
\format{A 802x48 response matrix}
\usage{
data(neoOpen)
}
\description{
A response matrix (n = 802) of NEO-PI-3's Openness to Experience
from Christensen, Cotter, & Silvia (2018).
}
\examples{
data("neoOpen")

}
\references{
Christensen, A. P., Cotter, K. N., & Silvia, P. J. (2018).
Reopening openness to experience: A network analysis of four openness to experience inventories.
\emph{Journal of Personality Assessment}, 1-15.
doi: \href{https://doi.org/10.1080/00223891.2018.1467428}{10.1080/00223891.2018.1467428}.
OSF: \href{https://osf.io/954a7/}{osf.io/954a7/}
}
\keyword{datasets}
