% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkToolbox--master.R
\docType{data}
\name{neoOpen}
\alias{neoOpen}
\title{NEO-PI-3 Openness to Experience Data}
\format{A 802x48 response matrix}
\usage{
data(neoOpen)
}
\description{
A response matrix (n = 802) of NEO-PI-3's Openness to Experience
from Christensen, Cotter, & Silvia (under review).
}
\examples{
data(neoOpen)
}
\references{
Christensen, A.P., Cotter, K.N., & Silvia, P.J. (under review).
Reopening Openness to Experience:
A network analysis of four Openness to Experience inventories.
\url{http://doi.org/10.17605/OSF.IO/954A7}
}
\keyword{datasets}
