% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkToolbox--master.R
\name{prepboot}
\alias{prepboot}
\title{Bootstrapped Network Preprocessing}
\usage{
prepboot(data, iter = 100, n = 100, network = "none", binary = FALSE)
}
\arguments{
\item{data}{A set of data}

\item{iter}{Number of bootstrap iterations. Defaults to 100 iterations}

\item{n}{Number of people in each bootstrapped sample. Defaults to 100}

\item{network}{Should the matrix be filtered? Defaults to "none" (no filter)}

\item{binary}{Is dataset dichotomous? Defaults to FALSE. Set TRUE if dataset is dichotomous (tetrachoric correlations are computed)}
}
\value{
If no filter is selected, then a matrix of the correlation standard deviations (larger values are smaller standard deviations; i.e., 1-sd).
If filter is selected, then a weighted adjacency matrix is produced with the correlations of the most stable connections in the network (i.e., smallest standard deviations)
}
\description{
Bootstraps the sample to identify the most stable correlations
}
\examples{
sdmat<-prepboot(data)
tmfgfilteredmat<-prepboot(data, network = "TMFG")
dichtomoussdmat<-prepboot(data, binary = TRUE)
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
