% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkToolbox--master.R
\name{randnet}
\alias{randnet}
\title{Generates a Random Network}
\usage{
randnet(nodes, edges)
}
\arguments{
\item{nodes}{Number of nodes in random network}

\item{edges}{Number of edges in random network}
}
\value{
Returns an adjacency matrix of a random network
}
\description{
Generates a random network
}
\examples{
rand <- randnet(10,27)
}
\references{
Rubinov, M., & Sporns, O. (2010). 
Complex network measures of brain connectivity: Uses and interpretations. 
\emph{Neuroimage}, \emph{52}(3), 1059-1069.
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
