% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkToolbox--master.R
\name{semnetmeas}
\alias{semnetmeas}
\title{Semantic Network Measures}
\usage{
semnetmeas(A, ...)
}
\arguments{
\item{A}{An adjacency matrix of network A}

\item{...}{Additional arguments for ASPL and CC (weighted)}
}
\value{
Returns a values for ASPL, CC, Q, and S
}
\description{
Computes the average shortest path length (ASPL),
clustering coefficient(CC),
modularity (Q),
and small-worldness (S; defaults to "rand" and 100 iterations)
}
\examples{
A<-TMFG(hex)$A

globmeas<-semnetmeas(A)
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
