% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkToolbox--master.R
\name{Strength}
\alias{Strength}
\title{Node Strength}
\usage{
Strength(A)
}
\arguments{
\item{A}{An adjacency matrix of network data}
}
\value{
A vector of strength values for each node in the network
}
\description{
Computes strength of each node in a network
}
\examples{
Str<-Strength(A)
}
\references{
Rubinov, M., & Sporns, O. (2010). 
Complex network measures of brain connectivity: Uses and interpretations. 
Neuroimage, 52(3), 1059-1069.
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
